# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchFilterPayload": {
        "description": "Payload model for optional search filters.",
        "properties": {
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optionales Genre-Filter",
            "title": "Genre"
          },
          "quality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optionale Qualit\u00e4tsanforderung (z. B. FLAC, 320kbps)",
            "title": "Quality"
          },
          "year": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optionales Jahr-Filter",
            "title": "Year"
          }
        },
        "title": "SearchFilterPayload",
        "type": "object"
      },
      "SearchFilters": {
        "description": "Filter options that may be applied to search results.",
        "properties": {
          "duration_ms": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "anyOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  {
                    "anyOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Inclusive duration range in milliseconds.",
            "title": "Duration Ms"
          },
          "explicit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Whether to include only explicit or clean tracks (Spotify only).",
            "title": "Explicit"
          },
          "genres": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "One or more genres to match (case-insensitive).",
            "title": "Genres"
          },
          "min_bitrate": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Minimum accepted bitrate in kbps.",
            "title": "Min Bitrate"
          },
          "preferred_formats": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Preferred audio file formats (used to boost relevance).",
            "title": "Preferred Formats"
          },
          "types": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "track",
                    "album",
                    "artist"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Restrict results to specific item types.",
            "title": "Types"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Restrict Soulseek matches to a specific username.",
            "title": "Username"
          },
          "year_range": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "anyOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  {
                    "anyOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Inclusive year range, expressed as [min, max].",
            "title": "Year Range"
          }
        },
        "title": "SearchFilters",
        "type": "object"
      },
...
      "SearchPagination": {
        "description": "Pagination options for the search endpoint.",
        "properties": {
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "default": 25,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Size",
            "type": "integer"
          }
        },
        "title": "SearchPagination",
        "type": "object"
      },
      "SearchResponse": {
        "description": "Response envelope for the unified search endpoint.",
        "properties": {
          "errors": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "items"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "SearchSort": {
        "description": "Sorting configuration for search results.",
        "properties": {
          "by": {
            "default": "relevance",
            "enum": [
              "relevance",
              "bitrate",
              "year",
              "duration"
            ],
            "title": "By",
            "type": "string"
          },
          "order": {
            "default": "desc",
            "enum": [
              "asc",
              "desc"
            ],
            "title": "Order",
            "type": "string"
          }
        },
        "title": "SearchSort",
        "type": "object"
      },
...
      "SpotifySearchResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "SpotifySearchResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__routers__sync_router__SearchRequest": {
        "description": "Request payload for the global search endpoint.",
        "properties": {
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchFilterPayload"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optionale Filter f\u00fcr Genre, Jahr und Qualit\u00e4t"
          },
          "query": {
            "description": "Freitext-Suchbegriff",
            "title": "Query",
            "type": "string"
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optionale Quellenauswahl (spotify, plex, soulseek)",
            "title": "Sources"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchRequest",
        "type": "object"
      },
      "app__schemas_search__SearchRequest": {
        "description": "Incoming payload for the unified search endpoint.",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/SearchFilters"
          },
          "pagination": {
            "$ref": "#/components/schemas/SearchPagination"
          },
          "query": {
            "description": "Free-text search query.",
            "title": "Query",
            "type": "string"
          },
          "sort": {
            "$ref": "#/components/schemas/SearchSort"
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "spotify",
                    "plex",
                    "soulseek"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Explicit set of sources to query. Defaults to all sources.",
            "title": "Sources"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Harmony Backend",
    "version": "1.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/search": {
      "post": {
        "description": "Perform a combined search across Spotify, Plex and Soulseek.",
        "operationId": "global_search_api_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__sync_router__SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Global Search Api Search Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Global Search",
        "tags": [
          "Sync",
          "Sync"
        ]
      }
    },
...
    "/search": {
      "post": {
        "description": "Aggregate search results across all configured music sources.",
        "operationId": "unified_search_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__schemas_search__SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unified Search",
        "tags": [
          "Search"
        ]
      }
    },
...
    "/spotify/search/artists": {
      "get": {
        "operationId": "search_artists_spotify_search_artists_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Query",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifySearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Artists",
        "tags": [
          "Spotify"
        ]
      }
    },
...
  }
}